moving the custom xrt devices to a seperate file (so they can be used by SXN)
